model-server/requirements.txt
flask
flask-cors
scikit-learn
joblib
numpy
gunicorn
pymupdf 
Pillow
opencv-python